メモの保存・読み込み処理をmemoStoreに移動
